credit-score-prediction/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bidict==0.22.1
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
curl-cffi==0.5.10
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
dill==0.3.8
distro==1.8.0
dnspython==2.4.2
duckduckgo_search==4.1.1
exceptiongroup==1.2.0
fastapi-socketio==0.0.10
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
imutils==0.5.4
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==3.1.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.2.10
langsmith==0.1.93
Lazify==0.4.0
libclang==16.0.6
loguru==0.7.2
lxml==5.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
mysqlclient==2.2.4
namex==0.0.7
nest-asyncio==1.5.8
oauthlib==3.2.2
onnxruntime==1.16.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
overrides==7.4.0
packaging==23.2
pandas-datareader==0.10.0
peewee==3.17.1
pillow==10.2.0
posthog==3.1.0
protobuf==4.25.1
psutil==6.0.0
pulsar-client==3.4.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyprojroot==0.3.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uptrace==1.22.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
watchdog==4.0.0
watchfiles==0.20.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
